Use `path.win32` from "path" over importing from "path/win32" (#199)
* Use `path.win32` from path over importing from `path/win32`

This is because `path/win32` was only added in Node.js v16. In Node.js
v10 to v14 only `path.win32` is available - which is still available in
v16, and v17, as well.

* Update CHANGELOG

* Run CommonJS tests on all supported Node.js versions

And all "normal" tests only on the development Node.js version. This is
necessary because the source code is written in ESM, but Node v10.13.0
does not support ESM, so running the "normal" tests is non-trivial.
Running the CommonJS test suite on every supported Node.js version will
at least catch faulty imports and, to an extends, faulty implementations
for older Node.js versions.

* Adopt CommonJS test suite for Node.js v10

See Node.js v10.x documentation for assert:
https://nodejs.org/dist/latest-v10.x/docs/api/assert.html