Add AVA as a devDependency
Add AVA [1] as a devDependency to replace mocha [2] as a testing
framework.

A couple of frameworks were considered, including Mocha, Jest, Tap, uvu,
and Tape. In the end AVA was chosen partly to try something new (I've
used Mocha, Jest, and uvu before), partly because it's newer and "more
modern", and partly because of the built-in Macros feature [3].

--
1. https://www.npmjs.com/package/ava
2. https://www.npmjs.com/package/mocha
3. ava/docs/01-writing-tests.md#reusing-test-logic-through-macros