Add initial property test suite with AVA
Add a first property test suite with ava(-fast-check) for the
`getDefaultShell` function of the win.js module. The tests are nearly
identical to the existing Mocha-based property tests for this function.

One improvement was made, the `process.env` generating custom
`fc.object` arbitrary was extracted into a helper module for 1) easier
re-use and 2) cleaner test code. This effect is already present in the
re-use of this custom arbitrary in the other property tests added here.

A custom script for running property tests with ava has been added as at
this point the new property tests do not yet replace the existing
property tests (due to lack of coverage). Once all property tests have
been rewritten the existing `test:property` script should be substituted
for the here-added `test:property-ava` script.