License fixtures with MPL-2.0
An interesting side effect of having the fixtures separately from the
tests where they're used is that they can be licensed differently. As
such, the fixtures will henceforth be licensed under the MPL-2.0 license
- simply put, requiring anyone making changes to it to make those
changes available under the same license. What this means for these
fixtures is that if someone discovers a new useful test case, it should
be made publicly available and so Shescape can benefit from that test
case.