Switch testing framework to AVA (#203)
Change(*) the testing framework from Mocha to AVA [1]. The main
motivation for this is just to try something new. A notable advantage of
AVA is better performance (primarily because of process isolation).
Other than that there's advantages (see the "Why AVA?" section of [1])
as well as disadvantage - but neither we're the reason for changing.

Mocha (*) is still kept for compatibility testing as AVA is a ES
module package, so can't be run in all Node.js version that need to be
used when running compatibility testing. This also means Mocha needs to
be kept below major version 10, per [2].

The goal of this rewrite was first and foremost just that, rewriting the
test. However, due to differences between Mocha and AVA this was not
necessarily straightforward and there's some quite significant changes.
Moreover, the rewrite was taken as an opportunity to improve things when
possible (e.g. improved re-use, remove meaningless or duplicate tests,
add missing tests). Hopefully, this did not result in any previously
covered test cases no longer being tested, but this isn't guaranteed.

The first notable change is the new test folder structure. First off,
it's no longer (mostly) flat. Instead, the first layer of the `test/`
folder now offers separation between different test categories, instead
of a semi-separation using file extension (e.g. `.prop.js` for property
tests). This consists of unit and e2e testing from the classic test
pyramid (integration testing is notably absent, one could argue the e2e
tests here are integration test. For now the project will just roll with
only e2e tests) as well as separation of property, compatibility, fuzz
and benchmark tests (note that fuzz and benchmark tests are outside the
scope of the framework changes, as they don't use Mocha/AVA, but had to
change in accordance with the folder structure changes as well as
changes to some common test utilities).

This change is partially required for the switch to AVA, but also makes
some sense more broadly speaking. For AVA, this is needed because test
names must be unique (otherwise you'll get an error), but test names do
include the location of the test file (relative to some common root).
This is also the reason for the second notably change, namely multiple
test files for a single file in the source code. This is true for unit,
e2e, compatibility, and property tests.

A related improvement is that there's now separate coverage reports for
each of the category, allowing for improved insights into each of them.
Utilizing these different coverage reports with Codecov is considered
future work. For additional information on these changes, see the
updates to the Contributing Guideline's "Testing" section.

Note that the compatibility category is a new addition. Previously the
"test:transpiled" category was used for compatibility testing. This new
category is an improvement of that, notably not requiring it to be run
for every OS due to the separation of it from the e2e tests.

Relatedly, there are some changes to the Continuous Integration configs
the test commands have changed. Notably, the job dependencies have been
updated due to improvements w.r.t. to compatibility testing.

Lastly, a call-out goes to the test fixtures. These are intentionally
licensed as MPL-2.0 instead of Unlicensed to encourage sharing new test
cases for shell escaping (as opposed to the test logic, which can be
used for other purposes as well and therefore has it's copyrights
waived).

--
1. https://github.com/avajs/ava/tree/714c474a49d41a53940b4a7516c1d1ed39b5eefe
2. https://github.com/ericcornelissen/shescape/pull/243