Add dedicated escaping logic for Dash
Based on the escaping logic for Bash, notably with escaping for tildes
(`~`) after equality signs (`=`) omitted. This isn't necessary based on
manual testing, and can be confirmed by running the existing fuzz target
on Dash (given that the corpus included in git contains various test
cases for the scenario, from [1]).

--
1. https://github.com/ericcornelissen/shescape/issues/169