Omit exhaustive default shell testing from e2e suites
Testing the behaviour of the default shell exhaustively (i.e. using all
examples provided by the fixtures) is prone to error as illustrated by
this branch of changes: the assumption was made that Bash is the default
shell, but in reality (at least for Ubuntu 20.04) the default shell
appears to be Dash (based on local testing and CI results). With that in
mind the default shell will no longer be exhaustively e2e tested.