Update e2e test macros
Update the e2e test macros to not test all fixtures on the default
shell. This is because the default shell may differen from system to
system, hence the e2e tests may fail on some systems depending on the
default shell.

Instead, the e2e tests now just verify that escaping an argument with
`{ shell: undefined }` doesn't throw an error.