Update e2e test macros, fix boolean shell value bug (#277)
* Update e2e test macros

Update the e2e test macros to not test all fixtures on the default
shell. This is because the default shell may differen from system to
system, hence the e2e tests may fail on some systems depending on the
default shell.

Instead, the e2e tests now just verify that escaping an argument with
`{ shell: undefined }` doesn't throw an error.

* Add e2e test for `{shell:false}` and `{shell:true}` options

* Fix error when `options.shell` is a boolean

Rather than checking if a shell was provided (i.e., the shell isn't
`undefined`), check if it's a string. If it's not a string (whether
`false`, `true`, or `undefined`), the default shell fallback should be
used - in line with child_process.

* Add unit tests for `options.shell` as a boolean

* Update JSDoc

To accurately reflect that `options.shell` may be a bool as well.

* Update CHANGELOG