Initialize vetting (#284)
* chore(deps): add unimported as a devDependency

Add unimported [1] as a devDependency, to be used for vetting the source
code of Shescape.

* chore: configure unimported

* chore(deps): add depcheck as a devDependency

Add depcheck [2] as a devDependency, to be used for vetting the
dependencies of Shescape.

* chore: configure depcheck

Configure depcheck to be explicit about the path of the ignore file
(`--ignore-path`) and include known false positives. Specifically
"rollup", which is used in the transpile command. Also configure it to
detect `.cjs` files and consider it during analysis. This is needed to
ensure dependencies used only in files with the `.cjs` extension (e.g.
"dotenv") are not reported as unused.

* chore: add "vet" command

Add a "vet" command for vetting the source code. Currently runs
unimported [1] and depcheck [2].

* ci: run `npm run vet` in the CI

--
1. https://www.npmjs.com/package/unimported
2. https://www.npmjs.com/package/depcheck