Update fuzzing with more fuzz targets (#275)
Add support for fuzzing targetting different child_process functions.
Where previously only `exec`/`execSync` were fuzz tested, this adds a
fuzz target for `spawn`/`spawnSync` (partial) and a skeleton for a fuzz
target for `execFile`/`execFileSync` (currently unsupported). More on
the incomplete support for these targets can be found below.

To support this, a script is created to run some pre-fuzz checks and
invoke jsfuzz [1] with the right fuzz target (based on an input provided
on the CLI). This script effectively replaces the old `prefuzz.js`
script.

The CI has been updated to run each of the (supported) fuzz targets. Now
for a shorter amount of time per target to keep the overall time
managable. Note, as this work resulted in some CI failures for fuzzing,
it was also discovered that the artifact names were invalid for all
`ubuntu-latests` shells due to those containing a `/`. Hence, this also
adds a "name" value to the fuzz matrix for usage in the artifact name.

For `spawn`/`spawnSync` only partial fuzzing support is provided due to
unexpected behaviour uncovered as part of this work. You can read more
about this at [2].

For `execFile`/`execFileSync` no fuzzing support is currently provided,
only a skeleton for fuzzing. This is in part due to [2], but also in
part due to a bug in the implementation of `execFileSync` - in
particular when `options.shell` is set to a non-falsy value, arguments
might not be passed on to underlying command correctly (depending on the
shell, but notably Bash and Dash exhibit this behaviour).

--
1. https://gitlab.com/gitlab-org/security-products/analyzers/fuzzers/jsfuzz
2. https://github.com/ericcornelissen/shescape/issues/286