Expand use of coverage in CI (#293)
* Run compat, property, and e2e tests with coverage in CI

* Push property and e2e test coverage into Codecov

Add a step to the properpty and e2e testing job of the CI to push the
coverage into Codecov. Use Codecov "flags" [1] to separate the property
and e2e and unit test coverage. For e2e tests, there's a flag for each
(tested, see matrix) operating system.

* Update Codecov comment configuration

Add the "flags" section [1] to the layout of the Codecov comment [2].

--
1. https://docs.codecov.com/docs/flags
2. https://docs.codecov.com/docs/pull-request-comments