Initialize an e2e test suite for usage with child_process
Create an initial e2e test suite for using Shescape with the Node.js
builtin package child_process [1]. This intent of this suite is to focus
on testing the interaction of Shescape with child_process when it's
actually used to invoke a command. In other words, it aims to create a
deterministic test suite of what the current fuzz tests test for.

There is one particular thing of note in this commit, the test macro for
`child_process.spawn`. It uses a test plan (`t.plan`) and custom
`Promise`:

- The latter is needed because AVA does not have support for callback
  testing (since v4 [2]), so a Promise that resolves when the spawned
  process `"close"`s is used instead (using `util.promisify` is very
  confusing in this case).
- The former is needed to ensure that the assertion inside the stdout
  callback is actually ran. Without the test plan the test would still
  succeed if there's no output on stdout.

--
1. https://nodejs.org/api/child_process.html
2. https://github.com/avajs/ava/releases/tag/v4.0.0