Create e2e test for usage with child_process' fork
Extend the e2e test suite for usage with child_process to include a test
for usage with the fork function.

As fork is exclusively for spawning Node.js modules, a Node.js module
has to be invoked rather than the `echo` command. Luckily, there's a
already a suitable(*) substitute for that in the repository, namely the
`echo.js` script used for fuzzing.

(*) However, the existing `echo.js` script for fuzzing only echos the
first argument. Hence, this upgrades the `echo.js` script to echo any
number of arguments, with a space in between each argument.

One note about the fork test is the force `silent: true` value on the
`options` object for the `cp.fork` call. This is necessary because
without it the output of the forked process will go directly to the
parents stdout (i.e. the terminal). This pollutes the test commands
output, but worse means the test cannot observe the output of the echo.
Hence, `silent` is always set to `true` in the `cp.fork` test macro.