Add recipe for Shescape usage with child_process' fork
Add two more recipes for usage of Shescape with the child_process
function fork. Despite not having the `shell` option available, which
means it is harder to misuse, it still has abuse potential. The provided
example shows this using the null character which, if not escaped, will
result in the remaining part of the argument containing it being
ignored.