Rename existing e2e test suite to integration
This name change follows from considerations of a new test suite (to be
added) that tests Shescape's interaction with child_process (and
potentially other similar packges). As that would be more end-to-end
than the current suite, this current suite has been renamed to
"integration" tests. This name was chosen as it tests how the different
modules of Shescape work together when integrated with eachother.