Create e2e test suite for usage with child_process (#297)
* Initialize an e2e test suite for usage with child_process

Create an initial e2e test suite for using Shescape with the Node.js
builtin package child_process [1]. This intent of this suite is to focus
on testing the interaction of Shescape with child_process when it's
actually used to invoke a command. In other words, it aims to create a
deterministic test suite of what the current fuzz tests test for.

There is one particular thing of note in this commit, the test macro for
`child_process.spawn`. It uses a test plan (`t.plan`) and custom
`Promise`:

- The latter is needed because AVA does not have support for callback
  testing (since v4 [2]), so a Promise that resolves when the spawned
  process `"close"`s is used instead (using `util.promisify` is very
  confusing in this case).
- The former is needed to ensure that the assertion inside the stdout
  callback is actually ran. Without the test plan the test would still
  succeed if there's no output on stdout.

Also, some tests are skipped. For all skipped tests this is due to known
bugs. The related issues are linked in the comment on the same line as
the skipped tests.

* Create commands to run e2e tests

* Run e2e tests in Continuous Integration

Update the test.yml workflow to include a job for running the e2e tests.
The e2e tests are run on multiple operating systems as behaviour between
operating system differs (both from Shescape's perspective as well as
shells).

The e2e tests are ran with coverage reporting enabled. The coverage
reports are uploaded to Codecov [3] with flags specific to the e2e
tests.

* Share echo.js script between fuzz and e2e tests

Move the constant defining the location of the echo.js script up to the
`test/_constants.cjs` file so it's easily accessible to both fuzz and
e2e tests. Also move the echo.js script out of the `test/fuzz` directory
and into the `test/` directory.

--
1. https://nodejs.org/api/child_process.html
2. https://github.com/avajs/ava/releases/tag/v4.0.0
3. https://about.codecov.io/