Update Node.js and npm versions for development (#291)
* Bump Node.js version for dev to 18, use lockfile version 3

* Update the Contributing Guidelines

Update the required version of Node.js and npm. lockfile version 3 is
only supported by npm v8 and up and prior to npm 8.1.2 there was a bug
with the `--lockfile-version` option inside of an `.npmrc` file.

The stylized names for "Node.js" and "npm" are also updated to be inline
with the official style (based on the website `<title>`s).

* Update compatibility testing job to use specific npm version

Add a step to the compatibility testing job to install a specific
version of npm that is known to be compatible with lockfile version 3.
This shouldn't affect the compatibility tests as those don't depend on
the npm version.

Based on CI logs, this is necessary for Node.js 10.13.0, 12, and 14. For
simplicity it happens for all compatability test jobs.

* Change `\u{0}` to `\u0000` in RegExps

The two are equivalent. But the former triggers a bug in Stryker.js
v6.0.1. For that reason, at least for now, `\u0000` is preferred.