Don't fail-fast in matrixed CI testing
Prevent failing fast and loosing the result of other test jobs in the
matrix for compatibility, integration, and e2e test. The results for
other Node.js versions (in the case of compatibility) and other
operating systems (in the case of integration and e2e) can be insightful
or helpful when trying to understand what went wrong.