Fix excessive escaping of single quotes on Unix systems
Fix the problem of excessive escaping of single quote characters on Unix
systems. The problem is that, when only escaping (w/o quoting), single
quote characters would be escaped unnecessarily as such escaping is only
necessary when the argument is wrapped in (single) quotes. To fix this,
the internals of Shescape now pass a `quoted` argument to the escape
functions. This argument, on Unix systems is used to determine what
should be escaped. (In the long run this option may be included in the
Shescape API, but that's a decision for later.)

The unit tests, fixtures, and integration test are updated in accordance
with these changes. The fixtures now include a new `expected` value for
when the new `quoted` option is true. As for most cases this is the same
as the `noInterpolation` case, users of the fixture default to that when
there's no `expected.quoted` value (for now). This change to fixtures is
the cause of the changes to integration tests.

The unit tests, in addition to being changed to accomodate the fixture
changes have some changes to the tests to accomodate the new internal
argument, mainly around asserting the value being passed is correct.