Fix unexpected fuzzing errors for some child_process functions
Reduce the amount of preparation done on arguments for various
child_process functions before using the argument to fuzz. The prep
is intended to change the argument in a way that does not affect
the command, but is necessary for sane results that is specific to
shells. This extra prep turns out to be unnecessary for the functions
for which it is being disabled.