Document use case for the `interpolation` option
Add a recipe for when/how to use the `interpolation` option for the
`escape` and `escapeAll` functions. While the recipe advices against
using it, it's still document so that at least there's a reference for
it's used. If there's no documentation about it at all, it's more likely
the option is misused.