Fix excessive escaping of quote characters (#310)
* Unskip fuzz targets detecting excessive escaping

* Unskip e2e tests related to excessive escaping

* Fix excessive escaping of single quotes on Unix systems

Fix the problem of excessive escaping of single quote characters on Unix
systems. The problem is that, when only escaping (w/o quoting), single
quote characters would be escaped unnecessarily as such escaping is only
necessary when the argument is wrapped in (single) quotes. To fix this,
the internals of Shescape now pass a `quoted` argument to the escape
functions. This argument, on Unix systems is used to determine what
should be escaped. (In the long run this option may be included in the
Shescape API, but that's a decision for later.)

The unit tests, fixtures, and integration test are updated in accordance
with these changes. The fixtures now include a new `expected` value for
when the new `quoted` option is true. As for most cases this is the same
as the `noInterpolation` case, users of the fixture default to that when
there's no `expected.quoted` value (for now). This change to fixtures is
the cause of the changes to integration tests.

The unit tests, in addition to being changed to accomodate the fixture
changes have some changes to the tests to accomodate the new internal
argument, mainly around asserting the value being passed is correct.

* Add CI fuzzing of `execFile` and `fork`

Reduce the fuzz time per child_process function to 5 minutes each,
keeping the total of 20 minutes of fuzzing.

* Update recipes, remove warning about known bug

* Test that the internal `quoted` option cannot be overridden

* Add child_process e2e tests for double quotes

Expand the e2e test suite for testing with the child_process package to
include test cases with double qoutes. This is relevant to Windows
systems, where double quotes may be excessively escaped.

* Reduce scope of "interpolation is omitted" unit test

Only test for the value of the `interpolation` argument, not (also) the
value of the `quoted` argument.

* Fix excessive escaping of double quotes on Windows systems

Fix the problem of excessive escaping of double quote characters on Win
systems. The problem is essentially identical to that of single quotes
on Unix, but in Windows it's double quotes because the `quote` function
wraps arguments in double quotes instead of single quotes.

* Fix unexpected fuzzing errors for some child_process functions

Reduce the amount of preparation done on arguments for various
child_process functions before using the argument to fuzz. The prep
is intended to change the argument in a way that does not affect
the command, but is necessary for sane results that is specific to
shells. This extra prep turns out to be unnecessary for the functions
for which it is being disabled.

* Remove fork from fuzz CI job

The `fork` fuzzing seems unreliable on Windows in the CI at this moment.
At this point it seems more important to merge the bug than have
continuous fuzzing for `fork` enabled.

* Update CHANGELOG