Document use case for the `interpolation` option (#312)
* Document use case for the `interpolation` option

Add a recipe for when/how to use the `interpolation` option for the
`escape` and `escapeAll` functions. While the recipe advices against
using it, it's still document so that at least there's a reference for
it's used. If there's no documentation about it at all, it's more likely
the option is misused.

* Remove unnecessary option in `shescape.quote` call

Per the `index.d.ts` file, `shescape.quote` does not accept the
`interpolation` option. Hence, it doesn't need to be specified for
fuzzing.