Test exact values for interpolation and quote
For the main module's `escapeShellArg` function, update the tests for
the `interpolation` and `quote` values to test for exact boolean values
rather than any "falsy" or "truthy" value. This allows the functions
being called to be simpler, as they only have to deal with boolean
values. The suite for `quoteShellArg` already tests the `undefined`
case as well as for exact boolean values.

Note: this test was created as a result of property tests being
developed for the main module's `escapeShellArg` and `quoteShellArg`
functions which uncovered that an `undefined` value for `interpolation`
to `escapeShellArg` isn't properly converted to a boolean value. Hence
the changes here should result in a failing unit test. Additionally, it
uncovered that `quoted` is currently relied upon being inherently falsy
(by virtue of being `undefined`) for `escapeShellArg`. It is preferred
for this to be explicitly `false` (as the changes to the test here
indicate), again to simplify the implementation of the escaping logic.