Remove redundant defaults for `escapeShellArg`
Per the mutation testing results (which replaced the removed object by
`{}`) the defaults object in `escapeShellArg` isn't necessary. This is
because of the change to `parseArgs`, which now explicitly sets either
`true` or `false` for the `interpolation` option based on the provided
value being truthy or falsy (resp.). Since omitting the options means
it's `undefined` (=falsy), it's automatically set to `false` already.

This simplifies things, especially in conjuction with the fact that the
options `{ interpolation: undefined }` should (clearly) behave identical
to `{ }`, but didn't bercause `{ interpolation: undefined }` overrides
`{ interpolation: false }` when merging objects.