Initialize the use of step-security/harden-runner
Start using step-security/harden-runner in the GitHub Actions CI by
setting it up in audit mode to detect outgoing connections so that it
can be configured later.

This Action adds an agent to the CI run that prevents outgoing
connections to any endpoint that's not whitelisted. In "audit" mode it
only records the outgoing connections, so that it can be configured to
only allow those previously detected in the future.