Configure step-security/harden-runner
Set egress-policy to block anything not in the allowlist and configure
the allowlist based on egress detected by the Action during audit. Also
add a name to the step-security/harden-runner job step.

The release workflow isn't configured yet because it's not ran in the
context of a Pull Request. The workflow can be configured after a
release has been created with the `egress-policy: audit` option.