Don't deal with newlines when fuzzing CMD
Avoid dealing with newline characters when fuzzing cmd.exe. It appears
that, when used programatically, everything after newline characters is
ignored when invoking a command with cmd.exe. For that reason, the fuzz
targets will continue not dealing with newlines, but only for cmd.exe.

Also add a new seed to the fuzz corpus, one that contains the `\r`
(carriage return) character, that is relevant to the cmd.exe behaviour.