Update fuzz/_common.cjs utilities to take shell as input
This enables it to be used both with and without a shell specified from
fuzz targets. The existing approach, where _common.cjs looks up the fuzz
shell itself, didn't always work because getting a prepared argument as
well as the expected output may depend on the shell being used. In the
event where the fuzz target wanted to test without specifying a shell,
but a shell was specified for fuzzing, this would go wrong.

The API for getting the prepared arg and the expected output has been
updated to improve readability, making more of it's arguments named by
using a paramater object as first argument.

All fuzz targets have been updated accordingly.