Start fuzzing whitespace with `interpolation:true`
Update the exec.test.cjs fuzz target to not strip out all whitespace and
instead let whitespace be in scope for fuzzing. This required some
changes to the prep and expected arg logic to accomodate the ways in
which CMD and PowerShell work with whitespace between arguments.

The additions to the fuzz corpus all relate to the way whitespace is
handled by CMD and PowerShell - they're for calibrating the fuzzing.

Unix shells were not tested as part of this commit.