Improve fuzz targets (#318)
* Stop not dealing with newlines for most fuzzing

Update most fuzz targets so that they now try out arguments containing
newlines. The exceptions to this are the fuzz targets with interpolation
set to true, these still ignore newlines (this is because in this
scenario newlines are not handled well yet, further work required); and
whenever fuzzing with cmd.exe (this is because cmd.exe does not handle
newlines well even when arguments are quoted, further work required) (a
item, "ec246439...", was added to the fuzz corpus for this).

* Align exec fuzz target with exec/execSync recipes

Expand the fuzz target for child_process.exec to include a check for
usage without having a shell specified (both with and without the
interpolation option set). Also, be more methodical about the checks,
following the "design" of the other fuzz targets.

* Fix incorrect argument preparation and expected output

Update the `prepareArg` and `getExpectedOutput` to take more inputs to
better understand how to respond. In particular, the inclusion of the
`shell` value ensures the correct return value is provided when a shell
**is** configured for fuzzing, but the particular fuzz check intents to
test without specifying a shell.

* Stop not dealing with (most) whitespace for most fuzzing interpolation

Update the fuzz checks that set interpolation to true to mostly start
fuzzing with whitespace, except for newline (see first paragraph). This
required some changes to the implementation (fixed in other branches
merged prior to this one). Multiple fuzz targets where added as a result
of this, namely:

- Related to whitespace between arguments
  - "377f0b03272b8e006a7ba3fb993328a87e7914075caae18209c7c8c1be205e14"
  - "38af29e913078ad07fc8d4156c2dfa3942e159bfcebe18e7e562ea3d82fce781"
  - "cb41f974ee87bf382bc82a4e275bea2ffb99d9bb8d502826f5b6d231250e20b7"
- Related to (shell-dependent) special characters after whitespace
  - "74cd192a0b85f46497a85e0b229630c55c306ce8904ca3d7e362c0ad81fa1088"
  - "ec246439c059f17f708aa39f76f88ec8d8eaee7ccca2f07d4c09ef57ae6503d8"
- Related to trimming whitespace at the start and end of the argumens
  - "bb9310963784ca202fd80f96f12a3cd13ee11f96a226d7cff5a921902a6b6324"
  - "f5520c691959b1105525b46c01e6055916ea029f439a5ccefd17b03742adef64"