Escape whitespace for PowerShell with `{interpolation:true}`
* Update test fixtures for PowerShell on Windows

... to include examples of whitespace and how it should be escaped. The
examples were obtained through fuzzing. Existing examples with relevant
characters have been updated as well.

The test macro used when running the fixtures has been updated to render
the various whitespace characters being added here. This is to ensure
unique test names.

* Escape necessary whitespace characters for PowerShell

For simplicity, this escapes all whitespace by converting it into spaces
(U+0020). The alternative, escaping whitespace with a backtick (`) seems
to be possible, but is fragile. Based on fuzzing, it doesn't always work
as expected, leading to unexpected changes in the input string. It is
deemed preferable to just replace all spaces as it's simpler to explain/
understand than seemingly random changes to the input.

In the future more granular escaping for (certain) whitespace characters
can be considered.