Escape whitespace for Bash, Dash, and Zsh with `{interpolation:true}`
* Update test fixtures for Bash, Dash, and Zsh on Unix

... to include examples of whitespace and how it should be escaped. The
examples were obtained based on the tests for PowerShell, as well as the
definition of the `\s` group for Regular Expression in JavaScript (the
former only includes `\u0085` that is not included in the latter).

All characters were manually testing using the script:

    import { execSync } from "node:child_process";
    import assert from "node:assert";
    import * as shescape from "shescape";

    const options = {
      interpolation: true,
      shell: "zsh",
      encoding: "utf8",
    };

    const whitespaceChars = [
      // Excluded because they will fail:
      // "\u0009", <- Converted to a normal space by Bash, Dash, and Zsh
      // "\u000A", <- Errors on Bash, Dash, and Zsh
      "\u000B", "\u000C", "\u000D", " ", "\u0085", "\u00A0", "\u1680",
      "\u2000", "\u2001", "\u2002", "\u2003", "\u2004", "\u2005",
      "\u2006", "\u2007", "\u2008", "\u2009", "\u200A", "\u2028",
      "\u2029", "\u202F", "\u205F", "\u3000", "\uFEFF",
    ];

    for (const char of whitespaceChars) {
      const userInput = `foo${char}bar`;

      const escapeOutput = execSync(
        `node test/_echo.js ${shescape.escape(userInput, options)}`,
        options
      );

      const quoteOutput = execSync(
        `node test/_echo.js ${shescape.quote(userInput, options)}`,
        options
      );

      assert.strictEqual(escapeOutput, quoteOutput);
    }

The test macro used when running the fixtures has been updated further
to render all whitespace characters correctly (to ensure unique test
names). This was updated to 1) output `\u{hhhh}` for readability (more
clearly demarking unicode characters from "normal" characters) and 2)
use unique character indices for all characters to more clearly
communicate the chosen order.

* Escape necessary whitespace characters for Bash, Dash, and Zsh

Only the newline character is necessary to escape for each shell
following the experiment outlined above.

In the future more granular escaping for whitespace characters can be
considered.