Update escaping of whitespace for CMD
* Add item to fuzz corpus

Add an item to fuzz corpus that triggers the bug on CMD where a trailing
forward slash in combination with a carriage return character results in
a double quote being outputted (similar to PowerShell's behaviour).

* Update test fixtures for CMD on Windows

So that it is now expected that carriage return characters are replaced
by a single space character.

Only escaping carriage return characters when `interpolation:true` was
manually tested through fuzzing, this showed that carriage return
characters always need to be escaped.

The following script was verified whether or not other line terminating
characters needed to be escaped as well:

    import { execSync } from "node:child_process";
    import assert from "node:assert";
    import fs from "node:fs";
    import * as shescape from "shescape";

    const tmp = fs.readFileSync(
      "./crash-" +
      "31ed7643aba69fe2d776af3aee587bb7899165af5ed3846c6f70327f2eec4713"
    ).toString();

    const newlineChars = [
      "\u000A",
      // "\u000D", <- '\r', does not work as we know
      "\u2028",
      "\u2029",
    ];

    for (const char of newlineChars) {
      const userInput = tmp.replace(/\r/, char);

      const escapeOutput = execSync(
        `node test/_echo.js ${shescape.quote(userInput)}`
      ).toString();

      assert.strictEqual(escapeOutput.trim(), tmp.trim());
    }

* Update escaping of whitespace for CMD on Windows

Note: the `_common.js` fuzz logic was updated accordingly to get the
correct expected output.