Escape curly bracket with line terminating characters for Bash/Dash
* Add item to fuzz corpus

Add an item to the fuzz corpus that triggers the bug on Bash/Dash where
brace expansion happens if there's a line terminating character within
the braces.

* Update test fixtures for Bash/Dash

...to include exmaples of the line terminating characters where the
curly brakcet should be escaped but currently isn't. The list of
characters is based on the experiments carried out for the tilde.

* Update escaping of curly bracket for Bash/Dash

Like the escaping of tildes for bash, use `[^]` to match _all_
characters, including newlines.

Note that Zsh does not (currently) need to be changed as it always has
curly braces escaped.

* Update the CHANGELOG