Improve testing and escaping of newlines (#332)
* Don't ignore newlines when fuzzing

Update the fuzzing logic to not ignore newline characters anymore. For
Bash, Dash, PowerShell, and Zsh this is reflected by the removal of

    arg = arg.replace(/[\n\r]+/g, "");

from the `exec.test.cjs` fuzz target. For CMD it is (additionally) the
removal of

    arg = arg.replace(/[\n\r]+/g, "");

in the `_common.cjs` fuzz logic.

The inclusion of newlines in input arguments required some changes to
the `getExpectedOutput` function to account for the individual shell's
handling of these characters, typically removing them from the start and
end of the input. (Note: not all changes to `getExpectedOutput` are due
to this.)

* Update the fuzz corpus

- `31ed764`: contains a trailing forward slash ('\\') character that
  was echod as a double quote ('"') character by CMD due to carriage
  return ('\r') characters not being escaped.
- `dece2a6`: contains a line feed ('\n') character and a tilde where the
  tilde should be escaped but wasn't for Bash.
- `f28febc`: contains a carriage return ('\r') character and a tilde
  where the tilde should be escaped but wasn't for Bash.
- `faf3c25`: contains a line feed ('\n') character inside of curly
  braces where the opening curly bracket should be escaped but wasn't
  for bash.

The remaining new items in the fuzz corpus (`036053b`, `71ef6c5`, and
`f1d97ce`) we're added based on fuzzing with whitespace for PowerShell.

* Update macro for executing fixtures

Update the macro for executing fixtures to properly display all the
whitespace characters as distinct characters. This is required because
otherwise test names will be considered duplicates.

Whitespace characters with common shorthands are display using that
shorthand (e.g. '\n' for line feed). Other characters will be displayed
by their unicode index (similar to the null character). This has been
normalized to the `\u{hhhh}` notation for consistency where the curly
braces are included to improve readability of the test title (clearly
demarcating the end of the unicode index from any characters that might
follow it).

* Escape newlines for PowerShell

For PowerShell only line feed ('\n') and carriage return ('\r')
characters need to be escaped, and only when `interpolation` is `true`.
This is supported by manual testing as well as fuzzing.

From manual testing (see [2]) it turned out that all character except
for line feed ('\n'), carriage return ('\r'), and '\uFEFF' are converted
to a normal space by PowerShell so it isn't necessary to escape them.
For '\uFEFF', it was found that it is interpreted as is and doesn't
cause any problems.

The test fixtures for PowerShell were updated to include examples where
whitespace characters need or need not be escaped.

* Escape newlines for CMD

For CMD only line feed ('\n') and carriage return ('\r') characters need
to be escaped, but those need to be escaped always. This is supported by
manual testing as well as fuzzing.

From manual testing (see [2]) it turned out that line feed ('\n')
characters cause CMD to omit the remaining input;  carriage return
('\r') could change the interpretation of certain characters; tab ('\t')
characters are converted to normal spaces; and all other whitespace
characters are interpreted as is.

The test fixtures for CMD were updated to include examples where
whitespace characters need or need not be escaped.

* Escape newlines for Bash, Dash, and Zsh

For the supported Unix shells only the line feed ('\n') character needs
to be escaped, and only when `interpolation` is `true`. This is
supported by manual testing as well as fuzzing.

From manual testing (see [2]) it turned out that line feed ('\n') and
characters cause the whatever follows it to be interpreted as a new
command; tab ('\t') characters are converted to normal spaces; and all
other whitespace characters are interpreted as is.

The test fixtures for Bash, Dash, and Zsh were updated to include
examples where whitespace characters need or need not be escaped.

* Improve escaping of tilde for Bash

For Bash escape the tilde ('~') character if a line terminating
character ('\n', '\r', '\u2028', or '\u2029') appears between and
"opening" equals sign ('=') and a "closing" colon-tilde (":~"). This
case was missed previously because the regular expression `/./` does not
indeed match all characters, but only those that are not line
terminating characters (even with the multiline flag). From [1]:

> Note that the m multiline flag doesn't change the dot behavior. So to
> match a pattern across multiple lines, the character class `[^]` can
> be used - it will match any character including newlines.

The test fixtures for Bash were updated to include examples where the
tilde character should have been escaped but wasn't.

* Improve escaping of curly braces for Bash

For Bash escape the opening curly brace ('}') character if a line
terminating character ('\n', '\r', '\u2028', or '\u2029') appears
between and opening curly brace ('{') and closing curly brace ('}').
Again, this case was missed previously because the regular expression
`/./` does match _all_ characters - instead `/[^]/` is used.

The test fixtures for Bash were updated to include examples where the
opening curly brace should have been escaped but wasn't.

* Update CHANGELOG

Include the fixed bugs in the CHANGELOG.

---

1. https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide-
   /Regular_Expressions/Character_Classes
2. Sample of a script used for manual testing:

    import { execSync } from "node:child_process";
    import assert from "node:assert";
    import * as shescape from "shescape";

    const options = {
      shell: "powershell.exe", // <-- set shell to test here
      interpolation: true,     // <-- change depending on test
      encoding: "utf8",
    };

    const whitespaceChars = [
      "\u0009", "\u000A", "\u000B", "\u000C", "\u000D", " ", "\u0085",
      "\u00A0", "\u1680", "\u2000", "\u2001", "\u2002", "\u2003",
      "\u2004", "\u2005", "\u2006", "\u2007", "\u2008", "\u2009",
      "\u200A", "\u2028", "\u2029", "\u202F", "\u205F", "\u3000",
      "\uFEFF",
    ];

    for (const char of whitespaceChars) {
      const userInput = `foo${char}bar`;

      const escapeOutput = execSync(
        `node test/_echo.js ${shescape.escape(userInput, options)}`,
        options
      );
      // Use this if you don't expect the escaped output to equal the
      // quote output.
      //assert.strictEqual(escapeOutput, "some string");

      const quoteOutput = execSync(
        `node test/_echo.js ${shescape.quote(userInput, options)}`,
        options
      );
      assert.strictEqual(escapeOutput, quoteOutput);
    }